Ensure FreeBSD Ruby install points at valid root CA cert
On FreeBSD the Ruby install's SSL points at `/etc/ssl/cert.pem` as the
default root CA cert. This is actually the location of an optional symlink
that the `ca_root_nss` port creates. We can't guarantee this optional
symlink will exist so we'll just point `SSL_CERT_FILE` at the actual
location of the root CA cert.