Update the url for ccache/rsync
The URL for ccache has changed from http -> https, additionally,
there is server side config that requires tlsv1 as the minimum
version. remote_file is offering sslv3 and the server is terminating
the connection.

Remove unsupported versions of ruby from travis

In exploring urls for ccache we found that the samba repo for rsync
was also having some problems. Favoring consistency, the base urls
for both ccache and rsycn were updated.