Add path for 7-zip
This is a workaround for a bug in either the 7-zip installer or the
cookbook.
Apparently 7-zip isn't respecting the install path on the command line,
and is installing itself to Program Files instead. But we set the path
to the directory we think it should be going in. Leaving the old
command in place just incase something fixes itself upstream.