Don't load mingw path for ruby
Devkit installs a thing which does this. It runs that whenever
rubygems needs to gem install something that has native extensions.

The issue with having this in the path is if you did something like
install a build toolchain that came before the devkit mingw path.
In this case, gem installing will not correctly use the devkit toolchain
because it only checks that Devkit/mingw is in the path, meaning the
toolchain we put in front of it will be used instead.