prefix all windows vars by OMNIBUS_
this is actually critically necessary for BUNDLER_VERSION
which collides horribly with the env var of the same name
in bundler 1.14.x which is used for bundler trampolining into
a specific version of bundler.

the value that this script is setting BUNDLER_VERSION to
is nonsensical for the new bundler code so it always errors
out, and consumers wouldn't want to be forced to be trampolined
back into the external bundler version all the time anyway.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>