Update Windows SDK installation method.
This is due to what appears to be a bug in the actual sdk setup executable.
When installing from a dos path (anything that has > 8 characters in the name),
the installer will truncate the download URL for some unknown reason, turning
<base_url>/Patches/<name> into <base_url>/hes/<name>. I have no idea what is
causing this but this code essentially forces a powershell invocation in the
current directory, which by default uses the correct "Long File" format.

Signed-off-by: Scott Hain <shain@chef.io>