Remove dependency on windows cookbook and disable FC121
FC121 alerts if you depend on cookbooks where the resources are included
in the chef-client now. We're actually using the homebrew recipe so I've
disabled that rule for now. I looked around and I can't figure out where
we're actually using the windows cookbook. I think it was once included
for the powershell_out logic, but that's been in Chef for a *long* time.
We'll still end up with in the dep tree since wix and seven_zip both
depend on it, but some day it will fade away and this is the first step

Signed-off-by: Tim Smith <tsmith@chef.io>