undo a change from previous cookstyle, and cookstyle
In commit ca4e42910a77c, the arguments to `File.join` were changed
from two arguments using the return from `build_user_home` and the
`load-omnnibus-toolchain.sh` file to using the return from
`build_user_home` sending an argument `load-omnibus-toolchain.sh`.
This causes an ArgumentError exception to be raised when attempting to
use the `omnibus_build` resource in an existing cookbook because
`build_user_home`'s method signature takes no arguments.

It's unclear whether the cookstyle of January 2019 caused an error on
this particular line of code, but I ran cookstyle 5.20.0 from Chef
Workstation 0.16.5 (latest stable release as of today) against this
change and it was not a finding. It did find an issue with the use of
`node.normal` in the `_compile.rb` recipe, so I fixed that unrelated
file in this change.

Signed-off-by: Joshua Timberman <joshua@chef.io>